Fix re-export of iscsi volume when using lioadm
This change ensures we check for existing chap credentials in the db
when re-exporting a volume. Without this change we overwrite the
existing credentials, resulting in an inaccessible volume.

This is a manual cut and paste of portions of change
Iea3d94e35a4ced4dafc1b61e2df6b075cf200577. Specifically, we add the
overridden _get_target_chap_auth for LioAdm, and add a context
argument to the function throughout.

It is not possible or desirable to backport the whole change for several
reasons:

* It is based on a refactor which is not on this branch
* It includes a refactor which is not relevant to this branch
* It includes more than 1 fix

Closes-bug: #1410566
Change-Id: I9859306b8beb0e1d8e3b3f6464c3f42e672e3949