RemoteFS: Fix the offline snapshot delete operation
The RemoteFS snapshot driver commits the wrong differencing disk
when performing an offline snapshot delete.

Because of this issue, the same snapshot will represent a different
point in time if the previous one is deleted. More information
about this situation is provided in the bug description.

This patch fixes the issue by commiting the right image in this
case.

Conflicts:
	cinder/volume/drivers/remotefs.py

Change-Id: Ifb5e840e5a918d727275a2c0be783e1fe045e798
Closes-bug: #1468445
(cherry picked from commit 0a360207112f5d21eb482feddfd88d5b6e10fdaa)