Handle FC LUN IDs greater 255 correctly on s390x architectures
----
NOTE(Stefan Amann): as of Liberty Cinder uses os-brick. That's the
reason why I cannot backport a valid Cinder patch from master.

os-brick contains this change with commit
75d1f38e3bfe487df77f7328ca6e86a15043d80b.
----

FC LUN IDs between hex 100 and hex ffff need to be formatted correctly
for the unit_add command. Otherwise, the volume will not be added to the
hypervisor configuration and the hypervisor will not be able to perform
I/O to it.

Closes-Bug: #1488233

Change-Id: I3910b5d7e0ae97e6b7f02c4b1c2be8ff8edd8dc7